+++ src/vagrant/storage/__init__.py
@@ -2,10 +2,13 @@
 
 from vagrant.storage.environments import Environment, EnvironmentManager
 from vagrant.storage.history import HistoryEntry, HistoryStorage
+from vagrant.storage.secrets import SecretStorage, is_secret_variable
 
 __all__ = [
     "HistoryEntry",
     "HistoryStorage",
     "Environment",
     "EnvironmentManager",
+    "SecretStorage",
+    "is_secret_variable",
 ]